refactor: DeleteTakenLecture UseCase Service 통합
