+++ src/main/java/com/plzgraduate/myongjigraduatebe/takenlecture/application/service/delete/DeleteTakenLectureService.java
@@ -4,8 +4,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.plzgraduate.myongjigraduatebe.completedcredit.application.usecase.GenerateOrModifyCompletedCreditUseCase;
 import com.plzgraduate.myongjigraduatebe.core.meta.UseCase;
-import com.plzgraduate.myongjigraduatebe.takenlecture.application.port.DeleteTakenLecturePort;
 import com.plzgraduate.myongjigraduatebe.takenlecture.application.usecase.delete.DeleteTakenLectureUseCase;
+import com.plzgraduate.myongjigraduatebe.takenlecture.application.port.DeleteTakenLecturePort;
 import com.plzgraduate.myongjigraduatebe.user.application.usecase.find.FindUserUseCase;
 import com.plzgraduate.myongjigraduatebe.user.domain.model.User;
 
@@ -14,11 +14,16 @@ import lombok.RequiredArgsConstructor;
 @UseCase
 @Transactional
 @RequiredArgsConstructor
-public class DeleteTakenLectureServiceById implements DeleteTakenLectureUseCase {
+class DeleteTakenLectureService implements DeleteTakenLectureUseCase {
 
 	private final FindUserUseCase findUserUseCase;
 	private final DeleteTakenLecturePort deleteTakenLecturePort;
 	private final GenerateOrModifyCompletedCreditUseCase generateOrModifyCompletedCreditUseCase;
+
+	@Override
+	public void deleteAllTakenLecturesByUser(User user) {
+		deleteTakenLecturePort.deleteAllTakenLecturesByUser(user);
+	}
 
 	@Override
 	public void deleteTakenLecture(Long userId, Long deletedTakenLectureId) {
