+++ src/main/java/com/plzgraduate/myongjigraduatebe/takenlecture/application/usecase/delete/DeleteTakenLectureUseCase.java
@@ -1,5 +1,10 @@
 package com.plzgraduate.myongjigraduatebe.takenlecture.application.usecase.delete;
 
+import com.plzgraduate.myongjigraduatebe.user.domain.model.User;
+
 public interface DeleteTakenLectureUseCase {
+	void deleteAllTakenLecturesByUser(User user);
+
 	void deleteTakenLecture(Long userId, Long deletedTakenLectureId);
+
 }
